requirements.txt
Django==4.2.5
pytest==7.4.3
djangorestframework==3.14.0
h11==0.14.0
outcome==1.3.0.post0
selenium==4.16.0
trio==0.23.1
trio-websocket==0.11.1
wsproto==1.2.0
elastic-transport==8.10.0
elasticsearch==8.11.0
elasticsearch-dsl==8.11.0
django-debug-toolbar==4.2.0
spacy~=3.7.2
django-elasticsearch-dsl==8.0
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0
PDFMiner==20191125
pyPDF==3.17.4
pycryptodome==3.19.1
unidecode==1.3.7
PyPDF2==3.0.1
